services/nlp/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
piper-tts>=1.2.0
mecab-python3>=1.0.6
jieba>=0.42.1
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.27.0
